requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
defusedxml==0.7.1
Django==4.2.7
django-bootstrap-icons==0.8.7
django-ckeditor==6.7.0
django-crispy-forms==2.1
django-js-asset==2.2.0
djangorestframework==3.14.0
idna==3.6
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
